+++ deid/dicom/pixels/detect.py
@@ -2,19 +2,27 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2016-2022, Vanessa Sochat"
 __license__ = "MIT"
 
+
 """
 detect.py: functions for pixel scrubbing
 """
 
-from pydicom import read_file
+from typing import Union, List, Optional
+
+from pydicom import read_file, FileDataset
 from pydicom.sequence import Sequence
 
 from deid.config import DeidRecipe
 from deid.dicom.filter import apply_filter
+from deid.dicom.types import DcmOrStr
 from deid.logger import bot
 
 
-def has_burned_pixels(dicom_files, force=True, deid=None):
+def has_burned_pixels(
+    dicom_files: Union[list[DcmOrStr], DcmOrStr],
+    force: bool = True,
+    deid: Optional[DeidRecipe] = None,
+):
     """has burned pixels is an entrypoint for has_burned_pixels_multi (for
     multiple images) or has_burned_pixels_single (for one detailed repor)
     We will use the MIRCTP criteria (see ref folder with the
@@ -34,7 +42,7 @@ def has_burned_pixels(dicom_files, force=True, deid=None):
     return _has_burned_pixels_single(dicom_files, force, deid)
 
 
-def _has_burned_pixels_multi(dicom_files, force, deid):
+def _has_burned_pixels_multi(dicom_files: List[Union[str, FileDataset]], force, deid):
     """return a summary dictionary with lists of clean, and then lookups
     for flagged images with reasons. The deid should be a deid recipe
     instantiated from deid.config.DeidRecipe. This function should not
@@ -58,7 +66,7 @@ def _has_burned_pixels_multi(dicom_files, force, deid):
     return decision
 
 
-def _has_burned_pixels_single(dicom_file, force, deid):
+def _has_burned_pixels_single(dicom_file: DcmOrStr, force: bool, deid):
 
     """has burned pixels single will evaluate one dicom file for burned in
     pixels based on 'filter' criteria in a deid. If deid is not provided,
@@ -102,7 +110,10 @@ def _has_burned_pixels_single(dicom_file, force, deid):
             ]
         }
     """
-    dicom = read_file(dicom_file, force=force)
+    if isinstance(dicom_file, FileDataset):
+        dicom = dicom_file
+    else:
+        dicom = read_file(dicom_file, force=force)
 
     # Return list with lookup as dicom_file
     results = []