+++ deid/dicom/utils.py
@@ -6,10 +6,14 @@ import os
 import tempfile
 import zipfile
 
+import pydicom
+from pydicom import FileDataset
+
 from deid.logger import bot
 from deid.utils import recursive_find
-
+from .types import DcmOrStr
 from .validate import validate_dicoms
+
 
 ################################################################################
 # Functions for Dicom files
@@ -101,3 +105,10 @@ def save_dicom(dicom, dicom_file, output_folder=None, overwrite=False):
     if dowrite:
         dicom.save_as(output_dicom)
     return output_dicom
+
+
+def load_dicom(dcm_file: DcmOrStr) -> FileDataset:
+    if isinstance(dcm_file, FileDataset):
+        return dcm_file
+    else:
+        return pydicom.read_file(dcm_file, force=True)